Hello, world
A few years of garbage history are folded into this commit, to only
keep some of the more useful remarks: in theory most of the details
and complexity *should* be characterised by tests which can be run
against github, though in reality most of it is likely in interactions
with the mergebot.

The codebase is pretty full of TODO, FIXME, unwrap, and todo! all of
which signal things to do. There's also a bunch of skipped tests
because I started validating some behaviour then it got too crazy or
too annoying to implement and I dropped it.

A few notable issues:

Fetch over dumb HTTP
====================

The negotiation system in the smart HTTP protocol is really not well
documented, and the original plan of "just send the entire network's
content" turned out to cause problematic divergences with
github (odoo/runbot#821).

So while eventually it would be a good idea to implement the smart
http protocols, for now the dumb protocol works out of the box (since
the client does all the hard work) and because we're working locally
with pretty small repositories it's unlikely to be any sort of
performance bottleneck, even with mitmproxy between DC and the SUT.

Commit message handling by github is a bit odd
==============================================

Apparently github trims commit messages when it returns them via the
API (and possibly on creation as well), even if the commit message
does actually genuinely contain leading and trailing whitespace. It is
essentially impossible to access those via the github API.

Furthermore, it will apparently replace non-printable characters by
the unicode replacement character in the HTML UI. In the API this
appears as unicode escapes (but might be valid verbatim, idk).

A third interesting component is that it's difficult to make git
create a non-UTF8 commit message: if a non-utf8 byte is provided, git
will assume the message is in latin1 and transcode from latin1 to utf8
before committing *that*, possibly garbling the message in the process
but always ending up with valid UTF8. Alternatively it's possible to
specify a `commitEncoding`, which will set the corresponding commit
header value.

Anyway it is possible to craft a mis-encoded commit message using
`hash-object` (`--literally` is not even necessary), the result of
which is that all of git, github-web, and github-api will decode with
replacement. So, you know, do that. Currently there's only 4 sites
where we need to decode commit messages so probably not worth any
complication, but we might want to newtype messages and bundle in the
conversion eventually (e.g. `impl From<BStr> for Message`).

Eventually we should also handle `commit.encoding` correctly and
transcode such message (assuming github does that, TBD but that would
make sense). According to the docs,

> other extended ASCII encodings are also supported [...] but not
> UTF-16/32, EBCDIC and CJK multi-byte encodings

So this seems to pretty much match the "legacy single byte encodings"
of the unicode standard. Meaning encoding_rs should work if we want to
implement it.

Tracing
=======

OTEL tracing can be enabled by setting the `--features otlp` and
setting `OTEL_SERVICE_NAME` and `OTEL_EXPORTER_OTLP_ENDPOINT` in the
environment (maybe, to verify, for the latter, I've not checked in a
while if it had a working default now but I'm pretty sure it didn't
work at one point). Specifically in my experience
`OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4317` works with
OpenTelemetry Desktop Viewer and Venator.

There have been some back and forths with this, some known issues:

- `err(Debug)` is a terrible idea as "error results" are pretty
  normal, *especially* in tests (as they're some of the most
  interesting edge cases). This causes the logs to be spammed with
  useless garbage e.g. screens after screens of

        dummy_central::github::repos::git error GHError { error: Unprocessable("Reference does not exist") ...

- It looks like events (logs) don't carry properly between tracing and
  opentelemetry, this should be investigated better but for now I just
  gave up and removed most of them (replacing some by span
  attributes).
- The opentelemetry bridge is kind-of an opinionated weirdo, would be
  a good idea to remove it and keep only the bits we have a use for
  it.
- OTEL (or at least OpenTelemetry Desktop Viewer and venator) don't
  have good support for linked spans (`follows_from`) unless that
  changed recently, so webhook delivery tracing was moved into a
  sub-span of their requester even though it's not really true...

Serde & validation
==================

A lot of github's validation can't be tested because while serde is
overall amazing it's really not great for specific openapi-style
validation (especially multifield).

To improve validation, it might be necessary to DOM-serialize, perform
procedural validation, then convert to the internal transfer
objects. Or switch to something which allows the sort of validation
necessary (possibly built over serde).

v4 API / GraphQL
================

As I don't really care for it, only the resolvers for which there is
no v3 API available are implemented. However adding new resolvers
should be fairly easy as I ended up adding `async-graphql` to
implement it.