Move fetch over from smart to dumb HTTP protocol
I'm nowhere near understanding the negotiation system in the smart
HTTP protocol, and the lazy man's fetch of packing and sending the
entire network causes significant behavioral differences with
github (see odoo/runbot#821).

While eventually it would be a good idea to re-add support for smart
fetch, for now using the dumb protocol to implement a precise fetch is
much easier. And since this is just reading and we're working over
localhost and the repositories we're working with are pretty small
it's unlikely to be any sort of *real* performance bottleneck.