chrono update deprecated `from_utc`
See chronotope/chrono#1175

The recommendation is to use `TimeZone::from_utc_datetime` but for our
use case I don't think this matters and this simple renaming is more
convenient. From my understanding this is the correct interpretation:
the timestamp is in UTC, and the offset is the one of the original
creator and mostly indicative.