Trim commit messages returned by API
After some minor testing, it's become clear that github trims commit
messages when it returns them via the API (and possibly on creation as
well), even if the commit message does actually genuinely contain
leading and trailing whitespace. It is essentially impossible to
access those via the github API.

An other interesting component is that git will apparently replace
non-printable characters by the unicode replacement character in the
HTML UI. In the API this appears as unicode escapes (but might be
valid verbatim, idk).

A third interesting component is that it's difficult to make git
create a non-UTF8 commit message: if a non-utf8 byte is provided, git
will assume the message is in latin1 and transcode from latin1 to
utf8 before committing *that*, possibly garbling the message in the
process but always ending up with valid UTF8. Alternatively it's
possible to specify a commitEncoding, which will set the corresponding
commit header value.

Anyway it is possible to craft a mis-encoded commit message using
`hash-object` (`--literally` is not even necessary), the result of
which is that all of git, github-web, and github-api will decode with
replacement. So, you know, do that. Currently there's only 4 sites
where we need to decode commit messages so probably not worth any
complication, but we might want to newtype messages and bundle in the
conversion eventually (e.g. `impl From<BStr> for Message`).

Eventually we should also handle `commit.encoding` correctly and
transcode such message (assuming github does that, TBD but that would
make sense). According to the docs,

> other extended ASCII encodings are also supported [...] but not
> UTF-16/32, EBCDIC and CJK multi-byte encodings

So this seems to pretty much match the "legacy single byte encodings"
of the unicode standard. Meaning encoding_rs should work if we want to
implement it.