Fix: error reporting of git/commit when the oid is invalid
- not a valid oid
- oid not found
- not a commit

should all map to a 422 for some weird ass reason.

Because the purpoted oid should be part of the message, and I don't
really want to type-erase, update `Error::Unprocessable` to use a Cow,
add a utility function `Error::unprocessable` for the old signature,
then... update all the existing callsites (which is a lot) to fix
things up.