Basic tracing improvements
Decided to start leveraging tracing a touch as it'd been configured a
while but not really *used*.

`#[instrument]` a bunch of root endpoints (in src/github.rs) so they
get logged out when called.

Configure the very commonly called `create_repository` to more finely
report what we would like, and remove the existing `span!` which is
now unnecessary.

Although mayhaps the span should be reintroduced (and the
`#[intrument]` removed) in order to provide a more amenable span
name-chain e.g. have all the spans for a repo be named after the
repository. With better reporting of repo names on the test side, and
storage of the tracing info in a somewhat durable format, this might
allow for a better reporting layout.

Then again, we could probably also just add `owner` and `name` fields
to the relevant spans (for most endpoints that would be
`#[intrument]`-able) and filter based on that. I assume tracing
collectors can expose this sort of things.