Fix a ton of unused things warnings
and reformat code while at it.

For the initial stuff I set a `#[allow(unused)]` so I wouldn't be
bothered, then I never removed it so there was a literal ton of unused
imports, locals, must_use, ... lying around.

Most of them were pretty easy to fix, so fixed them all I did. Leave
dead_code allowed for now as it covers unused struct fields, and
there'a s few of these lying around for structs I created but did not
fill in yet or something.