Tracing improvements: improve conf & instrument all endpoints
Instrumenting all endpoints provides the opportunity for better
visibility, although it's little more than an opportunity at this
point given the output is incredibly verbose already (despite close to
no non-span events being generated) and the test concurrency of the
mergebot means the output is unusably mixed.

Also improve the configuration to log out *span* information, may need
to be refined (as the currently the `close` event yields little more
than span in terms of usable data).

The tracing data currently suffers from a few major issues:

- events from unrelated traces are all mixed together
- while it's possible to relate requests to their source test via the
  repository they affect, this is neither convenient nor very reliable
- no integration whatsoever with the test or odoo logging, which makes
  it not very useful in case of test failure anyway

Thus I need to improve the following eventually:

- have *tests* generate trace ids, and send those trace ids when
  communicating with DC *or odoo*
- have the Odoo instance(s) better support and generate tracing data,
  and forward trace ids they (should) get from pytest
- have DC associate those trace ids with its own span
- finally put all the tracing data somewhere sensible, possibly on a
  per (session, test) basis, such that on failure it's possible to
  follow through the entire thing in one coherent sequence

At the moment, there's just no visibility in DC during tests, and even
in the logging generated by pytest the output from odoo and that of
the test harness is not correctly sequenced (because buffering and
stuff).