Soft-code users file for validation, and add bitwarden generator
Validation requires real github credentials with real accesses to
various accounts. Having a file with that in cleartext on-disk was
always dodgy and it's high time I do something more reliable.

Since Odoo has a bitwarden instance but (AFAIK) no secrets manager
account use private vault items to stash various bits of data and
secrets. Hopefully `mkusers` is sufficiently self-explanatory (most
likely it's not), support both `bw` and flatpak-provided `bw` directly
as the latter is intended to be aliased and scripts can't use
aliases. Obviously one could use a proxy script instead of an alias,
but it's easy enough to support the flatpak invocation so...